requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6
openai==1.59.7
aiosqlite==0.20.0
fuzzywuzzy==0.18.0
Levenshtein==0.26.1
mangum==0.19.0